% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_GSOD.R
\name{fetch_GSOD}
\alias{fetch_GSOD}
\title{Fetch GSOD Data and Subset Fields for Interpolation}
\usage{
fetch_GSOD(years = NULL, dsn = NULL)
}
\arguments{
\item{years}{A numeric vector of years of GSOD data to fetch for
interpolation. Defaults to current year.}

\item{dsn}{Optional. A filepath where resulting CSV files are to be saved on
local disk. If unspecified a tidy data frame is returned in the R session.}
}
\value{
List of data frames
}
\description{
This function can be wrapped in an \code{\link[base]{lapply}} function to
retrieve and save multiple years of GSOD data for interpolation, though it
may be used to retrieve GSOD data and interpolate files on-the-fly. See
\code{vignette("GRID")} for more details and examples.
}
\details{
This function will fetch GSOD data using
\code{\link[GSODR]{get_GSOD}} and save a CSV file containing only the
following fields to use in interpolating a global surface between 60 and -60
degrees latitude.
\describe{
\item{STNID}{A unique station id number that can be us ed to identify the
station and link with GSODR data for station metadata.}
\item{LON}{Longitude in decimal degrees.}
\item{LAT}{Latitude in decimal degrees.}
\item{ELEV_M_SRTM_90m}{Elevation in metres corrected for possible errors,
derived from the CGIAR-CSI SRTM 90m database (Jarvis et al. 2008).}
\item{YEAR}{The year (YYYY).}
\item{YDAY}{Sequential day of year.}
\item{TEMP}{Mean daily temperature in degrees C to tenths. Missing = NA.}
\item{MAX}{Maximum temperature reported during the day in Celsius to tenths.
 Missing = NA;}
\item{MIN}{Minimum temperature reported during the day in Celsius to tenths.
 Missing = NA;}
\item{RH}{Mean daily relative humidity.}
}
}
\examples{
\dontrun{

# Fetch one year of GSOD data
gsod_1998 <- fetch_GSOD(years = 1998)

# Fetch multiple years of GSOD data and save to disk
years <- as.list(seq(from = 1983, to = 2017, by = 1))
lapply(X = years, FUN = fetch_GSOD, dsn = "~/Data/GSOD")
}
}
\references{
Jarvis, A., Reuter, H. I., Nelson, A., Guevara, E. (2008)
Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM
90m Database (http://srtm.csi.cgiar.org)
}
