% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DEM.R
\name{get_DEM}
\alias{get_DEM}
\title{Get and Aggregate a Digital Elevation Model for Use in Interpolating GSOD Data}
\usage{
get_DEM(resolution = NULL, dsn = NULL)
}
\arguments{
\item{resolution}{Resolution to aggregate the digital elevation model to in
arc-degrees, e.g. 1 = 1 arc degree, .25 = one quarter arc degree. Valid
options are 1, .5 and .25 degrees.}

\item{dsn}{Optional. Directory where resulting DEM file is to be saved. If
unspecified a spatial object is returned in the R session. If a DEM exists,
it will be overwritten with the new one of the same resolution. If a new
resolution is specified, a new file will be created.}
}
\value{
A digital elevation model cropped to -60/60 degrees latitude and
aggregated by the requested factor
}
\description{
Getes a digital elevation model (DEM) from WorldClim data, crops and
aggregates to a larger spatial resolution and crops at -60/60 degrees
latitude for use with other weather data that provide rainfall, \emph{e.g.}
NASA - POWER at 1 arc degree. Mainly for use in crop modelling exercises.
}
\examples{

\dontrun{
# Get DEM and aggregate to 1 arc degree
DEM <- get_DEM(dsn = "~/Data/DEM")
}

}
