\name{getNames}
\alias{getNames}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ List the names of grobs on the display list }
\description{
  Returns a character vector containing the names of all
  top-level grobs on the display list.
}
\usage{
getNames()
}
\value{
  A character vector.
}
\author{ Paul Murrell }
\examples{
grid.grill()
getNames()
}
\keyword{ dplot }% at least one, from doc/KEYWORDS
