% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glint-package.R
\docType{package}
\name{glint}
\alias{glint}
\alias{glint-package}
\title{Global ('GSOD') Interpolated Temperatures}
\description{
Interpolates 'GSOD' temperature data as provided by the '\link[GSODR:GSODR]{GSODR::GSODR}'
package.
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
