\name{grid.clear}
\alias{grid.clear}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Clear a Grid Viewport}
\description{
  This function clears a Grid viewport by filling the viewport
  with the default background colour.

  This is NOT the same as erasing the viewport (see \code{grid.newpage}).
}
\usage{
grid.clear(clear.border=TRUE, vp=NULL)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{clear.border}{A logical value indicating whether to clear the
  viewport border as well as its interior.}
  \item{vp}{A Grid viewport object.}
}
\value{
  None.
}
\section{WARNING}{This function has been temporarily removed !}
\author{Paul Murrell}
\seealso{
  \code{\link{grid.newpage}}
}

\examples{
}
\keyword{ dplot }%-- one or more ...

