\name{grid.newpage}
\alias{grid.newpage}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Move to a New Page on a Grid Device}
\description{
  This function erases the current device or moves to a new page.
}
\usage{
grid.newpage(recording = TRUE)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{recording}{A logical value to indicate whether the new-page
    operation should be saved onto the Grid display list.}
}
\value{
  None.
}
\author{Paul Murrell}
\seealso{
  \link{Grid}
}

\examples{
}
\keyword{ dplot }%-- one or more ...
