\name{grid.plot.and.legend}
\alias{grid.plot.and.legend}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ A Simple Plot and Legend Demo }
\description{
  This function is just a wrapper for a simple demonstration of how
  a basic plot and legend can be drawn from scratch using grid.
}
\usage{
grid.plot.and.legend()
}
\author{ Paul Murrell }
\examples{
grid.plot.and.legend()
}
\keyword{ dplot }% at least one, from doc/KEYWORDS
