\name{grid.pretty}
\alias{grid.pretty}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{ Generate a Sensible Set of Breakpoints }
\description{
  Produces a pretty set of breakpoints within the range given.
}
\usage{
grid.pretty(range)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{range}{ A numeric vector }
}
\value{
  A numeric vector of breakpoints.
}
\author{ Paul Murrell }
\keyword{ dplot }% at least one, from doc/KEYWORDS
