\name{grid.refresh}
\alias{grid.refresh}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Refresh the current grid scene }
\description{
  Replays the current grid display list.
}
\usage{
grid.refresh()
}
\author{ Paul Murrell }
\keyword{ dplot }% at least one, from doc/KEYWORDS
