% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gsod.R
\name{interpolate_gsod}
\alias{interpolate_gsod}
\title{Interpolate GSOD Data to a Gridded Surface}
\usage{
interpolate_gsod(bz2_file, dem, dsn)
}
\arguments{
\item{bz2_file}{A bz2 CSV file generated by \link{fetch_gsod}.}

\item{dem}{Path to a raster readable digital elevation model file.}

\item{dsn}{where resulting CSV files are to be saved. Defaults
to user's "home" directory.}
}
\value{
GeoTIFF files of interpolated GSOD data at the spatial resolution of the
supplied digital elevation model.
}
\description{
This function is designed to be wrapped in an \code{\link[base]{lapply}}
function to retrieve multiple years of GSOD data for interpolation, though a
single year may be used.
}
\examples{
\dontrun{
dem <- raster::raster("~/Data/SRTM/SRTM_1deg.grd")
# Run the function
lapply(X = GSOD_list, FUN = interp, dem = dem)
}
}
