% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dem.R
\name{make_DEM}
\alias{make_DEM}
\title{Get and aggregate a digital elevation model for use in interpolating GSOD data}
\usage{
make_DEM(resolution = NULL, dsn = NULL)
}
\arguments{
\item{resolution}{Resolution to aggregate the digital elevation model to in
arc-degrees, e.g. 1 = 1 arc degree, 0.5 = one half arc degree, 0.25 = one
quarter arc degree. Valid options are \code{1}, \code{0.5} and \code{0.25}. Defaults to
\code{0.5}, the same as \acronym{NASA} \acronym{POWER}.}

\item{dsn}{Optional. Directory where resulting \acronym{DEM} file is to be
saved. If unspecified a spatial object is returned in the R session. If a
\acronym{DEM} exists, it will be overwritten with the new one of the same
resolution. If a new resolution is specified, a new file will be created.}
}
\value{
A \code{terra::SpatRaster()} object of a digital elevation model cropped
to -60/60 degrees latitude and aggregated by the requested factor and
optionally a data file written to local disk as a GeoTIFF object.
}
\description{
Fetches a digital elevation model \acronym{DEM} from 'WorldClim' data, crops
and aggregates to a larger spatial resolution and crops at -60/60 degrees
latitude for use with other weather data that provide rainfall, \emph{e.g.}
\acronym{NASA} \acronym{POWER} at 0.5 arc degrees or \acronym{TRMM} at 0.25
arc degrees. Mainly for use in crop modelling exercises.
}
\examples{

\donttest{
# Get DEM and aggregate to 0.5 arc degree, saving it to a local "~/Data/DEM"
# directory.

#DEM <- make_DEM(dsn = "~/Data/DEM")
}
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
