\name{unit.c}
\alias{unit.c}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Combine Unit Objects}
\description{
  This function produces a new unit object by combining the
  unit objects specified as arguments.
}
\usage{
unit.c(...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{\dots}{An arbitrary number of unit objects.}
}
\value{
  An object of class \code{unit}.
}
\author{Paul Murrell}
 \seealso{
   \code{\link{unit}}.
 }
\examples{}
\keyword{ dplot }%-- one or more ...
