\name{unit.length}
\alias{unit.length}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Length of a Unit Object}
\description{
  The length of a unit object is defined as the number of unit
  values in the unit object.
}
\usage{
unit.length(unit)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{unit}{A unit object.}
}
\value{
  An object of class \code{unit}.
}
\author{Paul Murrell}
\seealso{
  \code{\link{unit}}
}
\examples{
}
\keyword{ dplot }%-- one or more ...
