% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_DEM.R
\name{fetch_DEM}
\alias{fetch_DEM}
\title{Fetch and Aggregate a Digital Elevation Model for Use in Interpolating GSOD Data}
\usage{
fetch_DEM(resolution = NULL, dsn = NULL)
}
\arguments{
\item{resolution}{Resolution to aggregate the digital elevation model to in
arc-degrees, e.g. 1 = 1 arc degree, .25 = one quarter arc degree. Valid
options are 1, .5 and .25 degrees.}

\item{dsn}{Directory where resulting CSV files are to be saved. Defaults
to user's "home" directory.}
}
\value{
A digital elevation model cropped to -60/60 degrees latitude and
aggregated by the requested factor
}
\description{
Fetch and Aggregate a Digital Elevation Model for Use in Interpolating GSOD Data
}
\examples{

\dontrun{
# Fetch DEM and aggregate to 1 arc degree
fetch_DEM(dsn = "~/Data/DEM")
}

}
