% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa.r
\name{get_nasa}
\alias{get_nasa}
\title{Download NASA/POWER temperature and relative humidity variables and write to local disk}
\usage{
get_nasa(lon, lat, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
  stdate = "1983-1-1", endate = Sys.Date(), dsn = NULL, filename = NULL)
}
\arguments{
\item{vars}{Weather variables to download, Temperature, Minimum Temperature,
Maximum Temperature, Relative Humidity}

\item{stdate}{Starting date for download, defaults to 01/01/1983}

\item{endate}{End date for download, defaults to current date}

\item{dsn}{Local file path to write file out to}

\item{filename}{The filename for resulting file(s) to be written to disk}

\item{x}{Longitude in decimal degrees for cell to download}

\item{y}{Latitude in decimal degrees for cell to download}
}
\description{
Download NASA/POWER temperature and relative humidity variables and write to local disk
}
\examples{
\dontrun{
get_nasa(lon = -179.5, lat = 89.5)
}

}
\author{
Jorrel Khalil S. Aunario, jaunario@gmail.com
}
