% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glint-package.R
\docType{package}
\name{glint}
\alias{glint}
\alias{glint-package}
\title{Global ('GSOD') Interpolated Temperatures and Relative Humidity}
\description{
Interpolates 'GSOD' temperature and relative humidity data as provided by the
\pkg{GSODR} package.
}
\author{
Adam H. Sparks, \email{adamhsparks@gmail.com}
}
