
\name{grid-internal}
\alias{[.unit}
\alias{[.unit.arithmetic}
\alias{[.unit.list}
\alias{.GRID.STATE}
\alias{.grid.gpar.names}
\alias{.grid.loaded}
\alias{.grid.unit.list}
\alias{Ops.unit}
\alias{Summary.unit}
\alias{.unit}
\alias{.unit.arithmetic}
\alias{.unit.list}
\alias{absolute}
\alias{absolute.units}
\alias{absolute.units.unit}
\alias{absolute.units.unit.arithmetic}
\alias{absolute.units.unit.list}
\alias{as.character.unit}
\alias{as.character.unit.arithmetic}
\alias{as.character.unit.list}
\alias{clear.display.list}
\alias{clearpage}
\alias{col.spec}
\alias{common.draw.axis}
\alias{convert.gpar}
\alias{copy}
\alias{current.viewport}
\alias{demo.locator}
\alias{draw.all}
\alias{draw.details}
\alias{draw.details.arrows}
\alias{draw.details.circle}
\alias{draw.details.collection}
\alias{draw.details.default}
\alias{draw.details.frame}
\alias{draw.details.glist}
\alias{draw.details.lines}
\alias{draw.details.line.to}
\alias{draw.details.move.to}
\alias{draw.details.points}
\alias{draw.details.polygon}
\alias{draw.details.rect}
\alias{draw.details.segments}
\alias{draw.details.text}
\alias{draw.details.viewport}
\alias{draw.details.xaxis}
\alias{draw.details.yaxis}
\alias{draw.frame.child}
\alias{editDetails}
\alias{editDetails.default}
\alias{editDetails.frame}
\alias{editDetails.xaxis}
\alias{editDetails.yaxis}
\alias{get.gpar}
\alias{get.value}
\alias{get.value.default}
\alias{get.value.grob}
\alias{grid.Call}
\alias{grid.Call.graphics}
\alias{grid.height}
\alias{grid.init.viewport.stack}
\alias{grid.legend}
\alias{grid.multipanel}
\alias{grid.panel}
\alias{grid.prop.list}
\alias{grid.strip}
\alias{grid.top.level.vp}
\alias{grid.width}
\alias{height}
\alias{height.default}
\alias{height.details.default}
\alias{height.details.frame}
\alias{height.details.rect}
\alias{height.details.text}
\alias{height.details.viewport}
\alias{height.frame}
\alias{height.lines}
\alias{height.post}
\alias{height.post.details}
\alias{height.post.details.default}
\alias{height.pre}
\alias{height.pre.details}
\alias{height.pre.details.default}
\alias{height.text}
\alias{inc.display.list}
\alias{is.even}
\alias{is.gpar}
\alias{is.grob}
\alias{is.layout}
\alias{is.odd}
\alias{is.unit}
\alias{is.unit.arithmetic}
\alias{is.unit.list}
\alias{is.viewport}
\alias{justifyX}
\alias{justifyY}
\alias{layout.heights}
\alias{layout.ncol}
\alias{layout.nrow}
\alias{layoutRegion}
\alias{layout.respect}
\alias{layout.torture}
\alias{layout.widths}
\alias{make.xaxis.labels}
\alias{make.xaxis.major}
\alias{make.xaxis.ticks}
\alias{make.yaxis.labels}
\alias{make.yaxis.major}
\alias{make.yaxis.ticks}
\alias{mod.dims}
\alias{new.col}
\alias{new.row}
\alias{num.col.specs}
\alias{num.row.specs}
\alias{origin.bottom}
\alias{origin.left}
\alias{origin.right}
\alias{origin.top}
\alias{pop.saved.Rpars}
\alias{pop.saved.gpars}
\alias{pop.vp}
\alias{print.grob}
\alias{print.unit}
\alias{print.viewport}
\alias{push.saved.Rpars}
\alias{push.saved.gpars}
\alias{push.vp}
\alias{record}
\alias{record.default}
\alias{record.grob}
\alias{record.viewport}
\alias{recycle.data}
\alias{row.spec}
\alias{saved.pars}
\alias{set.gpar}
\alias{set.value.grob}
\alias{set.viewport}
\alias{stack.viewports}
\alias{swap.origin.horizontal}
\alias{swap.origin.vertical}
\alias{unit.arithmetic}
\alias{unit.arithmetic.rep}
\alias{unit.length.unit}
\alias{unit.length.unit.arithmetic}
\alias{unit.length.unit.list}
\alias{unit.list}
\alias{unit.list.c}
\alias{unit.list.from.list}
\alias{unit.list.rep}
\alias{unset.gpar}
\alias{unstack.viewports}
\alias{updateCol}
\alias{updateRow}
\alias{valid.data}
\alias{valid.just}
\alias{valid.layout}
\alias{valid.origin}
\alias{valid.pch}
\alias{valid.unit}
\alias{valid.units}
\alias{valid.viewport}
\alias{validGP}
\alias{viewport.layout}
\alias{viewport.layout<-}
\alias{viewport.transform}
\alias{width}
\alias{width.default}
\alias{width.details.default}
\alias{width.details.frame}
\alias{width.details.rect}
\alias{width.details.text}
\alias{width.details.viewport}
\alias{width.frame}
\alias{width.lines}
\alias{width.post}
\alias{width.post.details}
\alias{width.post.details.default}
\alias{width.pre}
\alias{width.pre.details}
\alias{width.pre.details.default}
\alias{width.text}
\title{Internal Grid Functions}
\description{
 Internal Grid functions
}
\Usage{
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
