% File src/library/grid/man/grid-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{grid-package}
\alias{grid-package}
\docType{package}
\title{
  The Grid Graphics Package
}
\description{
  A rewrite of the graphics layout capabilities, plus some
  support for interaction.
}
\details{
This package contains a graphics system which supplements
old-style S graphics (see the \pkg{graphics} package).

Further information is available in the following
\link{vignettes}:
\tabular{ll}{
\code{displaylist} \tab Display Lists in grid (source, pdf)\cr
\code{frame} \tab Frames and packing grobs (source, pdf)\cr
\code{grid} \tab Introduction to grid (source, pdf)\cr
\code{grobs} \tab Working with grid grobs (source, pdf)\cr
\code{interactive} \tab Editing grid Graphics (source, pdf)\cr
\code{locndimn} \tab Locations versus Dimensions (source, pdf)\cr
\code{moveline} \tab Demonstrating move-to and line-to (source, pdf)\cr
\code{nonfinite} \tab Non-finite values (source, pdf)\cr
\code{plotexample} \tab Writing grid Code (source, pdf)\cr
\code{rotated} \tab Rotated Viewports (source, pdf)\cr
\code{saveload} \tab Persistent representations (source, pdf)\cr
\code{sharing} \tab Modifying multiple grobs simultaneously (source, pdf)\cr
\code{viewports} \tab Working with viewports (source, pdf)\cr
}

For a complete list of functions with individual help pages,
use \code{library(help="grid")}.
}
\author{

Paul Murrell \email{paul@stat.auckland.ac.nz}

Maintainer: R Core Team \email{R-core@r-project.org}

}
\references{
  Murrell, P. (2005) \emph{R Graphics}. Chapman & Hall/CRC Press.
}
\keyword{ package }

