\name{grid-package}
\alias{grid-package}
\docType{package}
\title{
The Grid Graphics Package
}
\description{
A rewrite of the graphics layout capabilities, plus some
support for interaction.
}
\details{
This package contains a new graphics system, which replaces
(or extends) \link[graphics:graphics-package]{old-style S graphics}.

Further information is available in the following \link[utils:vignette]{vignettes}:
\tabular{ll}{
\code{displaylist} \tab Display Lists in grid (source, pdf)\cr
\code{frame} \tab Frames and packing grobs (source, pdf)\cr
\code{grid} \tab Introduction to grid (source, pdf)\cr
\code{grobs} \tab Working with grid grobs (source, pdf)\cr
\code{interactive} \tab Editing grid Graphics (source, pdf)\cr
\code{locndimn} \tab Locations versus Dimensions (source, pdf)\cr
\code{moveline} \tab Demonstrating move-to and line-to (source, pdf)\cr
\code{nonfinite} \tab Non-finite values (source, pdf)\cr
\code{plotexample} \tab Writing grid Code (source, pdf)\cr
\code{rotated} \tab Rotated Viewports (source, pdf)\cr
\code{saveload} \tab Persistent representations (source, pdf)\cr
\code{sharing} \tab Modifying multiple grobs simultaneously (source, pdf)\cr
\code{viewports} \tab Working with viewports (source, pdf)\cr
}

For a complete
list of functions, use \code{library(help="grid")}.
}
\author{

Paul Murrell \email{paul@stat.auckland.ac.nz}

Maintainer: R Core Team \email{R-core@r-project.org}

}
\keyword{ package }

