\name{grid.copy}
\alias{grid.copy}
\title{Make a Copy of a Grid Graphical Object}
\description{
  This function is redundant and will disappear in future versions.
}
\usage{
grid.copy(grob)
}
\arguments{
  \item{grob}{A grob object.}
}
\value{
  A copy of the grob object.
}
\author{Paul Murrell}
\seealso{
  \code{\link{grid.grob}}.
}
\keyword{dplot}
