% File src/library/grid/man/grid.copy.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{grid.copy}
\alias{grid.copy}
\title{Make a Copy of a Grid Graphical Object}
\description{
  This function is redundant and will disappear in future versions.
}
\usage{
grid.copy(grob)
}
\arguments{
  \item{grob}{A grob object.}
}
\value{
  A copy of the grob object.
}
\author{Paul Murrell}
\seealso{
  \code{\link{grid.grob}}.
}
\keyword{dplot}
