% File src/library/grid/man/grid.plot.and.legend.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{grid.plot.and.legend}
\alias{grid.plot.and.legend}
\title{ A Simple Plot and Legend Demo }
\description{
  This function is just a wrapper for a simple demonstration of how
  a basic plot and legend can be drawn from scratch using grid.
}
\usage{
grid.plot.and.legend()
}
\author{ Paul Murrell }
\examples{
grid.plot.and.legend()
}
\keyword{dplot}
