\name{grid.pretty}
\alias{grid.pretty}
\title{ Generate a Sensible Set of Breakpoints }
\description{
  Produces a pretty set of breakpoints within the range given.
}
\usage{
grid.pretty(range)
}
\arguments{
  \item{range}{ A numeric vector }
}
\value{
  A numeric vector of breakpoints.
}
\author{ Paul Murrell }
\keyword{dplot}
