% File src/library/grid/man/grid.refresh.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{grid.refresh}
\alias{grid.refresh}
\title{ Refresh the current grid scene }
\description{
  Replays the current grid display list.
}
\usage{
grid.refresh()
}
\author{ Paul Murrell }
\keyword{ dplot }
