% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gsod.R
\name{interpolate_GSOD}
\alias{interpolate_GSOD}
\title{Interpolate GSOD Data to a Gridded Surface}
\usage{
interpolate_GSOD(file_list = NULL, dem = NULL, dsn = NULL,
  vars = NULL, cores = NULL)
}
\arguments{
\item{file_list}{A \code{[base::list()]} of data frames or CSV files of
GSOD data created by \code{make_GSOD_set()}.}

\item{dem}{Digital elevation model that has been fetched and processed using
\code{get_DEM()}.}

\item{dsn}{Optional. Directory where resulting GeoTIFF files are to be saved.}

\item{vars}{Weather variables to interpolate. Possible values are
\code{TEMP}, \code{MAX}, \code{MIN} and \code{RH}. Defaults to \code{TEMP}.}

\item{cores}{Number of cores to use for parallel processing. Defaults to 1 on
Windows OS or if not otherwise specified.}
}
\value{
A \code{[raster::stack()]} of daily interpolated weather variables.
}
\description{
This function is designed to be wrapped in an \code{[base::lapply()]}
function to process multiple years of GSOD data for interpolation, though a
single year may be used.
}
\examples{
\dontrun{
# Get and aggregate the raster digital elevation model
dem <- get_DEM()

# Create a list of GSOD files
files <- list.files("~/Data/GSOD", full.names = TRUE)

# Run the function for MAX and MIN temperature on a *nix system using 2 cores
GRID <- lapply(X = files, FUN = interpolate_GSOD, dem = dem,
dsn = "~/Cache/GTiff", vars = c("MAX", "MIN"), cores = 2)

# Run the function for MAX and MIN temperature using 1 core, will work on Win
GRID <- lapply(X = files, FUN = interpolate_GSOD, dem = dem, vars = "MAX")
}

}
\author{
\email{adamhsparks@gmail.com}
}
