% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gsod.R
\name{interpolate_GSOD}
\alias{interpolate_GSOD}
\title{Interpolate GSOD Data to a Gridded Surface}
\usage{
interpolate_GSOD(gsod = NULL, dem = NULL, dsn = NULL, vars = NULL)
}
\arguments{
\item{gsod}{A \code{\link[base]{list}} of data frames or CSV files of GSOD
data created by \link{fetch_GSOD}.}

\item{dem}{Digital elevation model that has been fetched and processed using
\code{\link{fetch_DEM}}.}

\item{dsn}{Optional. Directory where resulting GeoTIFF files are to be saved.}

\item{vars}{Weather variables to interpolate. Possible values are,
\code{TEMP}, \code{MAX}, \code{MIN}, \code{RH}. Defaults to \code{TEMP}.}
}
\value{
\code{List} of \code{\link[raster]{stack}} objects by year and variable
}
\description{
This function is designed to be wrapped in an \code{\link[base]{lapply}}
function to process multiple years of GSOD data for interpolation, though a
single year may be used.
}
\examples{
\dontrun{
# Create a list of GSOD files
gsod_list <- list.files("~/Data/gsod", full.names = TRUE)

# Fetch and aggregate the raster digital elevation model
dem <- fetch_DEM()

# Run the function for MAX and MIN temperature
lapply(X = gsod_list, FUN = interpolate_GSOD,
       dem = dem, vars = c("MAX", "MIN"))
}

}
