% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gsod.R
\name{interpolate_GSOD}
\alias{interpolate_GSOD}
\title{Interpolate GSOD Data to a Gridded Surface}
\usage{
interpolate_GSOD(GSOD = NULL, dem = NULL, dsn = NULL, vars = NULL,
  cores = NULL)
}
\arguments{
\item{GSOD}{A \code{\link[base]{list}} of data frames or CSV files of GSOD
data created by \link{fetch_GSOD}.}

\item{dem}{Digital elevation model that has been fetched and processed using
\code{\link{fetch_DEM}}.}

\item{dsn}{Optional. Directory where resulting GeoTIFF files are to be saved.}

\item{vars}{Weather variables to interpolate. Possible values are,
\code{TEMP}, \code{MAX}, \code{MIN}, \code{RH}. Defaults to \code{TEMP}.}

\item{cores}{Number of cores to use for parallel processing. Defaults to 1 on
Windows OS or if not specified.}
}
\value{
\code{List} of \code{\link[raster]{stack}} objects by year and variable
}
\description{
This function is designed to be wrapped in an \code{\link[base]{lapply}}
function to process multiple years of GSOD data for interpolation, though a
single year may be used.
}
\examples{
\dontrun{
# Create a list of GSOD files
file_list <- list.files("~/Data/GSOD", full.names = TRUE)

# Fetch and aggregate the raster digital elevation model
dem <- fetch_DEM()

# Run the function for MAX and MIN temperature on a *nix system using 2 cores

library(parallel)
GRID <- lapply(X = file_list, FUN = interpolate_GSOD, dem = dem,
dsn = "~/Cache/GTiff", vars = c("MAX", "MIN"), cores = 2)

# Run the function for MAX and MIN temperature using 1 core, will work on Win

library(parallel)
GRID <- lapply(X = file_list, FUN = interpolate_GSOD, dem = dem,
dsn = "~/Cache/GTiff", vars = c("MAX", "MIN"))
}

}
