% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gsod.R
\name{interpolate_gsod}
\alias{interpolate_gsod}
\title{Interpolate GSOD Data to a Gridded Surface}
\usage{
interpolate_gsod(gsod = NULL, dem = NULL, dsn = NULL, vars = NULL)
}
\arguments{
\item{gsod}{A \code{\link[base]{list}} of data frames or CSV files of GSOD
data created by \link{fetch_gsod}.}

\item{dem}{Path to a raster readable digital elevation model file that has
been fetched and processed using \code{\link{get_DEM}}.}

\item{dsn}{Directory where resulting GeoTIFF files are to be saved. Defaults
to user's "home" directory.}

\item{vars}{Weather variables to interpolate. Possible values are,
\code{TEMP}, \code{MAX}, \code{MIN}, \code{RH}. Defaults to \code{TEMP}.}

\item{cores}{Number of processing cores to use in parallel. Will default to 1
on Windows OS, and on all others if unspecified.}
}
\value{
GeoTIFF files of interpolated GSOD data at the spatial resolution of the
supplied digital elevation model.
}
\description{
This function is designed to be wrapped in an \code{\link[base]{lapply}}
function to process multiple years of GSOD data for interpolation, though a
single year may be used.
}
\examples{
\dontrun{
# Create a list of GSOD files
gsod_list <- list.files("~/Data/gsod", full.names = TRUE)

# Load the raster digital elevation model
dem <- raster::raster("~/Data/SRTM/SRTM_1deg.grd")

# Run the function for MAX and MIN temperature
lapply(X = gsod_list, FUN = interp, dem = dem, vars = c("MAX", "MIN"))
}

}
