\name{stringWidth}
\alias{stringWidth}
\alias{stringHeight}
\title{Create a Unit Describing the Width of a String}
\description{
  These functions create a unit object describing the width or height of
  a string.
}
\usage{
stringWidth(string)
stringHeight(string)
}
\arguments{
  \item{string}{A character vector.}
}
\value{
  A unit object. 
}
\author{Paul Murrell}
\seealso{
  \code{\link{unit}} and
  \code{\link{grobWidth}}
}
\keyword{dplot}
