\name{unit.c}
\alias{unit.c}
\title{Combine Unit Objects}
\description{
  This function produces a new unit object by combining the
  unit objects specified as arguments.
}
\usage{
unit.c(...)
}
\arguments{
  \item{\dots}{An arbitrary number of unit objects.}
}
\value{
  An object of class \code{unit}.
}
\author{Paul Murrell}
\seealso{
  \code{\link{unit}}.
}
\keyword{dplot}
